Rework & Correct `Field` test
