+++ oxidation/libparsec/crates/serialization_format/src/protocol/parser/field.rs
@@ -39,7 +39,7 @@ impl Field {
         &self,
         visibility: syn::Visibility,
         types: &HashMap<String, String>,
-    ) -> syn::ExprType {
+    ) -> syn::Field {
         let (rename, name) = self.quote_name();
         let (ty, serde_skip) = self.quote_type(types);
 
@@ -56,14 +56,18 @@ impl Field {
             attrs.push(syn::parse_quote!(#[serde(default = #default)]));
         }
         if let Some(introduced_in) = self.introduced_in {
-            let introduced_text = format!("Introduced in API-{introduced_in}");
+            let introduced_text = format!("Introduced in `API-{introduced_in}`.");
             attrs.push(syn::parse_quote!(#[doc = #introduced_text]));
         }
 
-        syn::parse_quote! {
-            #(#attrs)*
-            #visibility #name: #ty
-        }
+        syn::parse::Parser::parse2(
+            syn::Field::parse_named,
+            quote::quote! {
+                #(#attrs)*
+                #visibility #name: #ty
+            },
+        )
+        .unwrap()
     }
 
     fn quote_name(&self) -> (Option<&'static str>, syn::Ident) {
@@ -92,17 +96,134 @@ impl Field {
     }
 }
 
+pub fn quote_fields(fields: &[Field], types: &HashMap<String, String>) -> Vec<syn::Field> {
+    fields
+        .iter()
+        .map(|field| field.quote(visibility_public(), types))
+        .collect()
+}
+
+fn visibility_public() -> syn::Visibility {
+    syn::parse_quote!(pub)
+}
+
 #[cfg(test)]
-#[rstest::rstest]
-#[case::basic_field(
-    r#"{"name": "Foo", "type": "String"}"#,
-    Field { name: "Foo".to_string() , ty: "String".to_string(), introduced_in: None, default: None }
-)]
-#[case::field_introduced_in(
-    r#"{"name": "Bar", "type": "Boolean", "introduced_in": "5.2"}"#,
-    Field { name: "Bar".to_string(), ty: "Boolean".to_string(), introduced_in: Some("5.2".try_into().unwrap()), default: None}
-)]
-fn field(#[case] input: &str, #[case] expected: Field) {
-    let field = serde_json::from_str::<Field>(input).expect("Got error on valid data");
-    assert_eq!(field, expected)
+mod test {
+    use std::collections::HashMap;
+
+    use ::quote::{quote, ToTokens};
+    use pretty_assertions::assert_eq;
+    use proc_macro2::TokenStream;
+    use rstest::rstest;
+
+    use super::{Field, MajorMinorVersion};
+
+    #[rstest]
+    #[case::basic_field(
+        r#"{"name": "Foo", "type": "String"}"#,
+        Field { name: "Foo".to_string() , ty: "String".to_string(), introduced_in: None, default: None }
+    )]
+    #[case::field_introduced_in(
+        r#"{"name": "Bar", "type": "Boolean", "introduced_in": "5.2"}"#,
+        Field { name: "Bar".to_string(), ty: "Boolean".to_string(), introduced_in: Some("5.2".try_into().unwrap()), default: None}
+    )]
+    fn deserialization(#[case] input: &str, #[case] expected: Field) {
+        let field = serde_json::from_str::<Field>(input).expect("Got error on valid data");
+        assert_eq!(field, expected)
+    }
+
+    #[rstest]
+    #[case::public(
+        syn::parse_quote!(pub),
+        Field {
+            name: "foo".to_string(),
+            ty: "Integer".to_string(),
+            introduced_in: None,
+            default: None,
+        },
+        quote! {
+            #[serde_as(as = "_")]
+            pub foo: i64
+        }
+    )]
+    #[case::public_with_introduced(
+        syn::parse_quote!(pub),
+        Field {
+            name: "foo".to_string(),
+            ty: "Integer".to_string(),
+            introduced_in: Some(MajorMinorVersion { major: 4, minor: 1}),
+            default: None,
+        },
+        quote! {
+            #[serde_as(as = "libparsec_types::Maybe<_>")]
+            #[serde(default, skip_serializing_if = "libparsec_types::Maybe::is_absent")]
+            #[doc = "Introduced in `API-4.1`."]
+            pub foo: libparsec_types::Maybe<i64>
+        }
+    )]
+    #[case::public_with_default(
+        syn::parse_quote!(pub),
+        Field {
+            name: "foo".to_string(),
+            ty: "Integer".to_string(),
+            introduced_in: None,
+            default: Some("42".to_string())
+        },
+        quote! {
+            #[serde_as(as = "_")]
+            #[serde(default = "42")]
+            pub foo: i64
+        }
+    )]
+    #[case::private(
+        syn::Visibility::Inherited,
+        Field {
+            name: "foo".to_string(),
+            ty: "Integer".to_string(),
+            introduced_in: None,
+            default: None,
+        },
+        quote! {
+            #[serde_as(as = "_")]
+            foo: i64
+        }
+    )]
+    #[case::private_with_introduced(
+        syn::Visibility::Inherited,
+        Field {
+            name: "foo".to_string(),
+            ty: "Integer".to_string(),
+            introduced_in: Some(MajorMinorVersion { major: 4, minor: 1}),
+            default: None,
+        },
+        quote! {
+            #[serde_as(as = "libparsec_types::Maybe<_>")]
+            #[serde(default, skip_serializing_if = "libparsec_types::Maybe::is_absent")]
+            #[doc = "Introduced in `API-4.1`."]
+            foo: libparsec_types::Maybe<i64>
+        }
+    )]
+    #[case::private_with_default(
+        syn::Visibility::Inherited,
+        Field {
+            name: "foo".to_string(),
+            ty: "Integer".to_string(),
+            introduced_in: None,
+            default: Some("42".to_string())
+        },
+        quote! {
+            #[serde_as(as = "_")]
+            #[serde(default = "42")]
+            foo: i64
+        }
+    )]
+    fn quote(#[case] vis: syn::Visibility, #[case] field: Field, #[case] expected: TokenStream) {
+        assert_eq!(
+            field
+                .quote(vis, &HashMap::new())
+                .into_token_stream()
+                .to_string(),
+            expected.to_string()
+        )
+    }
 }
