--- oxidation/libparsec/crates/serialization_format/src/tests/mod.rs
@@ -1,1 +1,0 @@
-mod quote_field;
